This are the beginnings of "simple mode". The idea is to have a simple view that doesn't have a concept of a Sickbeard module, for example - it uses all available data and just gives you and at-a-glance view of what has been recently added to your library (episodes and movies), and what episodes are going to be downloaded soon.
The idea is simplicity - it needs to look good, and give you just enough information to be useful (as this is mostly what I find myself using Maraschino for these days).

So far, it:

* displays 25 most recent episodes, styled nicely
* has placeholder columns for recent movies and upcoming episodes

To do:

* needs a fix to make the columns 100% in height and width
* needs some sort of Maraschino branding
* some very carefully thought out extras MAY be added at a later date, as long as it doesn't detract from the idea behind simple mode

Also, this is just an experiment when I was bored one night. It might not go anywhere.